cube + mvp + shape creator working
